add(feat/mqtt_service): .env for manage pswd / user / db createions / etc ...
