+++ src/main.ts
@@ -1,26 +1,13 @@
-import { INestApplication } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 import { devicesName } from './@types/fakeData';
 import { AppModule } from './app.module';
 import { generateFakeData } from './fakeData';
-import { description, version, name } from './../package.json';
 import { MQTTClient } from './mqtt/main';
+import { config } from 'dotenv';
+import { generateSwaggerDoc } from './swagger/swagger';
 
-/**
- * Generate swagger documenation
- * @param app Nest application
- */
-function generateSwaggerDoc(app: INestApplication): void {
-  const config = new DocumentBuilder()
-    .setTitle(name)
-    .setDescription(description)
-    .setVersion(version)
-    .addTag('api')
-    .build();
-  const document = SwaggerModule.createDocument(app, config);
-  SwaggerModule.setup('/swagger', app, document);
-}
+// Get env variables
+config();
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
@@ -29,12 +16,15 @@ async function bootstrap() {
 }
 
 if (process.env.STATUS === 'data') {
+  console.log('🚀 => Starting fake data...');
   for (const item of devicesName) {
     console.log(`=> Generating data for ${item.name}`);
     generateFakeData(1000, item);
   }
 } else if (process.env.STATUS === 'mqtt') {
+  console.log('🚀 => Starting MQTT client...');
   MQTTClient();
 } else {
+  console.log('🚀 => Starting server...');
   bootstrap();
 }
