+++ src/swagger/swagger.ts
@@ -1,0 +1,18 @@
+import { description, version, name } from './../../package.json';
+import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+import { INestApplication } from '@nestjs/common';
+
+/**
+ * Generate swagger documenation
+ * @param app Nest application
+ */
+export function generateSwaggerDoc(app: INestApplication): void {
+  const config = new DocumentBuilder()
+    .setTitle(name)
+    .setDescription(description)
+    .setVersion(version)
+    .addTag('api')
+    .build();
+  const document = SwaggerModule.createDocument(app, config);
+  SwaggerModule.setup('/swagger', app, document);
+}
